Upgrade dependencies and enhance specs.
Specs have been added to make the behavior of various usecases clearer.

In particular, the discussion in issue #11 should be made clearer with these spec changes. It is possible to ensure that tokens are scoped only to valid applications but this logic must be implemented as part of the `resource_owner_from_assertion` block.